fix: build error 수정
