+++ public/sw.js
@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const a=(a,n)=>(a=new URL(a+".js",n).href,s[a]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=a,e.onload=s,document.head.appendChild(e)}else e=a,importScripts(a),s()})).then((()=>{let e=s[a];if(!e)throw new Error(`Module ${a} didn’t register its module`);return e})));self.define=(n,i)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(s[c])return;let t={};const r=e=>a(e,c),o={module:{uri:c},exports:t,require:r};s[c]=Promise.all(n.map((e=>o[e]||r(e)))).then((e=>(i(...e),t)))}}define(["./workbox-e9849328"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"9e6c87e45a337f2670f9497b09a3776b"},{url:"/_next/static/chunks/144-5bb18d67640c7830.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/150d7f15-865e17574520807a.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/324.af65c5d1450aeab3.js",revision:"af65c5d1450aeab3"},{url:"/_next/static/chunks/337.756c8ed51c02b009.js",revision:"756c8ed51c02b009"},{url:"/_next/static/chunks/522-f3a90a96c10ca0e9.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/958-1fd1827977171caa.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/app/(home)/layout-6413ae1f4d2ee546.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/app/(home)/page-cc6aa10c890d2540.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/app/_not-found/page-fa44d575b494a900.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/app/layout-9101cb0395905941.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/app/login/page-a48f5cabf6d3fd85.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/framework-9b59ef6292ca80a3.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/main-17dbae30abe20dfb.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/main-app-8e63eb8d6001df9a.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/pages/_app-3caa9b0cf541d8c2.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/pages/_error-bc5f2b9116ac1ff1.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-054bd49bf12122af.js",revision:"ck4rj9hGfkWAaIZA1eAFb"},{url:"/_next/static/ck4rj9hGfkWAaIZA1eAFb/_buildManifest.js",revision:"85cc43490933360b0001f9f94bb52121"},{url:"/_next/static/ck4rj9hGfkWAaIZA1eAFb/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/css/a13d696f9caecca2.css",revision:"a13d696f9caecca2"},{url:"/icons/Arrow_Left.svg",revision:"08671dc557e1334588a73f262568f707"},{url:"/icons/Headphones.svg",revision:"650a5f7c532869b327869ab98df90e51"},{url:"/icons/Home.svg",revision:"4443225541e87e4138c873ec66967f67"},{url:"/icons/Shopping_Cart.svg",revision:"3a996a9d33d69b3ffac6dda95f9e7b5b"},{url:"/icons/User.svg",revision:"a0a1db6ba5cee52612172fe8b81a1401"},{url:"/img/Logo.png",revision:"fbf57ddbd716eadb019d257fecc9e94c"},{url:"/img/logo-192x192.png",revision:"fc39bf8a3420105fc6d392341048ba6d"},{url:"/img/logo-512x512.png",revision:"0ad62119e22f52383a69b1a83caba548"},{url:"/manifest.json",revision:"bd2d120e9566d381ab66b3991a20dbee"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:a,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let e,s={};const n=(n,t)=>(n=new URL(n+".js",t).href,s[n]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=n,e.onload=s,document.head.appendChild(e)}else e=n,importScripts(n),s()})).then((()=>{let e=s[n];if(!e)throw new Error(`Module ${n} didn’t register its module`);return e})));self.define=(t,i)=>{const a=e||("document"in self?document.currentScript.src:"")||location.href;if(s[a])return;let c={};const r=e=>n(e,a),o={module:{uri:a},exports:c,require:r};s[a]=Promise.all(t.map((e=>o[e]||r(e)))).then((e=>(i(...e),c)))}}define(["./workbox-e9849328"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"1e50387c22cd2ea2fc1eddf4f2b1ab16"},{url:"/_next/static/chunks/1-390821ec04f025e6.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/30-16624142bf5111a7.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/323-955bf41548e85da9.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/47edcb22-450264588a2cb1b5.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/482-ab7a7a1947ef9ea4.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/651-d292af2030f1586c.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/(home)/layout-d6564bd667bffe1e.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/(home)/page-6401600518b243cb.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/_not-found/page-bc04bff95e91ad3a.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/address-check/page-f9941550f5a96d83.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/api/naver/coordinates/route-400d18b072b1e439.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/api/naver/directions/route-bea992dd3527b2b7.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/layout-6e66b4b00a17bea4.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/login/page-4b65b3bcaea91997.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/app/login/step/page-726a8a5b5868f9f5.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/framework-289b5a20f31bf2b1.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/main-app-8682f331d99f10ce.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/main-c10c17f9b3d6e73e.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/pages/_app-7e5d4a60281b6427.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/pages/_error-77fa64d81664a8ec.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/chunks/polyfills-42372ed130431b0a.js",revision:"846118c33b2c0e922d7b3a7676f81f6f"},{url:"/_next/static/chunks/webpack-2f965e6ffa95848c.js",revision:"usLDAAtLLc54dsLSrzOlT"},{url:"/_next/static/css/3b5d2d92d5118830.css",revision:"3b5d2d92d5118830"},{url:"/_next/static/usLDAAtLLc54dsLSrzOlT/_buildManifest.js",revision:"e1db0779cbef5f7969f7736d0695d069"},{url:"/_next/static/usLDAAtLLc54dsLSrzOlT/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/icons/Arrow_Left.svg",revision:"d7adaa5ceadedcb1117c32fecc9486b7"},{url:"/icons/Headphones.svg",revision:"517fcc09c27e89a427db493cf4974896"},{url:"/icons/Shopping_Cart-active.svg",revision:"6f842f523ecabff234e64af894322117"},{url:"/icons/Shopping_Cart-inactive.svg",revision:"e7e18fedf253050eb043a984c788f1cf"},{url:"/icons/home-active.svg",revision:"f3cab85b38b8a345a74ba1803b8dc366"},{url:"/icons/home-inactive.svg",revision:"fa4dff6ab9a74f408a3f1b969f1df485"},{url:"/icons/user-active.svg",revision:"08de244efa3a443b5d9e2218a99ff3b1"},{url:"/icons/user-inactive.svg",revision:"e2a3105e917621a991bee1134e70398d"},{url:"/img/Checker.png",revision:"ef487bb5ee990730400c0a0a721ab10d"},{url:"/img/Logo.png",revision:"fbf57ddbd716eadb019d257fecc9e94c"},{url:"/img/logo-192x192.png",revision:"fc39bf8a3420105fc6d392341048ba6d"},{url:"/img/logo-512x512.png",revision:"0ad62119e22f52383a69b1a83caba548"},{url:"/manifest.json",revision:"d8ecc0eb149350f547e7ff82ebff2dee"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:n,state:t})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
